server/requirements.txt
bcrypt==3.1.4
Django==2.0.1
django-axes==4.0.1
django-brutebuster==0.1.8
django-cors-headers==2.1.0
django-formtools==2.1
django-polls==0.2.1
django-registration==2.4.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
flake8==3.5.0
flake8-docstrings==1.3.0
virtualenv==15.0.1
whitenoise==3.3.1
gunicorn==19.7.1
psycopg2==2.7.3.2
